feature_source/requirements.txt
rich>=13.7.0
py-clob-client>=0.18.0
websockets>=12.0
aiohttp>=3.9.0
numpy>=1.26.0
scipy>=1.11.0
python-dotenv>=1.0.0
web3>=6.0.0
fastapi>=0.109.0
uvicorn>=0.27.0